Extracting zipper to hole map conversion methods into a trait.
